fix migration

+++ database/migrations/2021_04_27_072155_create_case_management_regiment_form.php
@@ -13,22 +13,24 @@ class CreateCaseManagementRegimentForm extends Migration
      */
     public function up()
     {
-        Schema::create('case_management_regiment_form', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('form_id');
-            $table->foreign('form_id')->references('id')->on('tb_mac_forms')->onDelete('cascade');
-
-            $table->string('month_of_treatment')->default('');
-            $table->string('current_type_of_case')->default('');
-            $table->string('current_regiment')->default('');
-            $table->string('reason_case_management_presentation')->default('');
-            $table->string('current_weight')->default('');
-            $table->string('suggested_weight')->default('');
-            $table->string('updated_type_of_case')->default('');
-            $table->string('suggested_regimen')->default('');
-            $table->string('if_for_empiric_treatment')->default('');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('case_management_regiment_form')) {
+            Schema::create('case_management_regiment_form', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('form_id');
+                $table->foreign('form_id')->references('id')->on('tb_mac_forms')->onDelete('cascade');
+    
+                $table->string('month_of_treatment')->default('');
+                $table->string('current_type_of_case')->default('');
+                $table->string('current_regiment')->default('');
+                $table->string('reason_case_management_presentation')->default('');
+                $table->string('current_weight')->default('');
+                $table->string('suggested_weight')->default('');
+                $table->string('updated_type_of_case')->default('');
+                $table->string('suggested_regimen')->default('');
+                $table->string('if_for_empiric_treatment')->default('');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
